+++ OpenttdDiscord.Infrastructure.Tests/OpenttdDiscord.Infrastructure.Tests.csproj
@@ -44,4 +44,8 @@
 
 
 
+    <ItemGroup>
+      <Folder Include="Servers\Runners\" />
+    </ItemGroup>
+
 </Project>
